requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
python-multipart==0.0.9
 
# LangChain
langchain==0.2.0
langchain-google-genai==1.0.4
 
# Gemini
google-generativeai==0.5.4
 
# NLTK
nltk==3.8.1
 
# File parsing
PyPDF2==3.0.1
python-docx==1.1.0
 
# MySQL
mysql-connector-python==8.3.0
 
# Config
pydantic-settings==2.2.1
 
# Static files
aiofiles==23.2.1